requirements.txt
# requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
PyPDF2==3.0.1
sentence-transformers==3.1.1
langchain==0.3.0
faiss-cpu==1.8.0
google-generativeai==0.8.1
python-dotenv==1.0.1
langchain-huggingface